suppressed all warnings
suppressed all warnings

+++ Tutorial11_GraphicsEngine_RenderingCube/VertexBuffer.cpp
@@ -16,7 +16,7 @@ VertexBuffer::VertexBuffer():m_layout(0),m_buffer(0)
 {
 }
 
-bool VertexBuffer::load(void* list_vertices,UINT size_vertex,UINT size_list,void*shader_byte_code,UINT size_byte_shader)
+bool VertexBuffer::load(void* list_vertices,UINT size_vertex,UINT size_list,void*shader_byte_code,size_t size_byte_shader)
 {
 	if (m_buffer)m_buffer->Release();
 	if (m_layout)m_layout->Release();
