+++ Tutorial14_GraphicsEngine_Camera/DeviceContext.cpp
@@ -61,8 +61,8 @@ void DeviceContext::drawTriangleStrip(UINT vertex_count, UINT start_vertex_index
 void DeviceContext::setViewportSize(UINT width, UINT height)
 {
 	D3D11_VIEWPORT vp = {};
-	vp.Width = width;
-	vp.Height = height;
+	vp.Width = (FLOAT)width;
+	vp.Height = (FLOAT)height;
 	vp.MinDepth = 0.0f;
 	vp.MaxDepth = 1.0f;
 	m_device_context->RSSetViewports(1, &vp);
