+++ Tutorial17_GraphicsEngine_Texturing/Texture.cpp
@@ -15,7 +15,7 @@ Texture::Texture(const wchar_t* full_path): Resource(full_path)
 		D3D11_SHADER_RESOURCE_VIEW_DESC desc = {};
 		desc.Format = image_data.GetMetadata().format;
 		desc.ViewDimension = D3D11_SRV_DIMENSION_TEXTURE2D;
-		desc.Texture2D.MipLevels = image_data.GetMetadata().mipLevels;
+		desc.Texture2D.MipLevels = (UINT)image_data.GetMetadata().mipLevels;
 		desc.Texture2D.MostDetailedMip = 0;
 
 		GraphicsEngine::get()->getRenderSystem()->m_d3d_device->CreateShaderResourceView(m_texture, &desc,
