+++ Tutorial5_GraphicsEngine_DrawingTriangle/DeviceContext.cpp
@@ -40,8 +40,8 @@ void DeviceContext::drawTriangleList(UINT vertex_count, UINT start_vertex_index)
 void DeviceContext::setViewportSize(UINT width, UINT height)
 {
 	D3D11_VIEWPORT vp = {};
-	vp.Width = width;
-	vp.Height = height;
+	vp.Width = (FLOAT)width;
+	vp.Height = (FLOAT)height;
 	vp.MinDepth = 0.0f;
 	vp.MaxDepth = 1.0f;
 	m_device_context->RSSetViewports(1, &vp);
